# Introduce JsString, JsBoolean etc and support efficient string concatenation