--- a/Jint.Tests.Ecma/Properties/AssemblyInfo.cs
+++ b/Jint.Tests.Ecma/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Xunit;