--- a/Jint.Tests/ReplaceCulture.cs
+++ b/Jint.Tests/ReplaceCulture.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Globalization;
 using System.Reflection;
-using System.Threading;
 using Xunit.Sdk;
 
 namespace Jint.Tests