--- a/Jint.Tests/Runtime/Domain/Person.cs
+++ b/Jint.Tests/Runtime/Domain/Person.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Jint.Tests.Runtime.Domain
+﻿namespace Jint.Tests.Runtime.Domain
 {
     public class Person : IPerson
     {