--- a/Jint.Tests/Runtime/Domain/Shape.cs
+++ b/Jint.Tests/Runtime/Domain/Shape.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Jint.Tests.Runtime.Domain;
 
 namespace Shapes