--- a/Jint.Tests/Runtime/InteropTests.cs
+++ b/Jint.Tests/Runtime/InteropTests.cs
@@ -412,7 +412,7 @@ namespace Jint.Tests.Runtime
             var o = new
             {
                 x = new JsValue(1),
-                y = new JsValue("string"),
+                y = new JsString("string"),
             };
 
             _engine.SetValue("o", o);
@@ -427,7 +427,7 @@ namespace Jint.Tests.Runtime
         public void PocosCanReturnObjectInstanceDirectly()
         {
             var x = new ObjectInstance(_engine) { Extensible = true};
-            x.Put("foo", new JsValue("bar"), false);
+            x.Put("foo", new JsString("bar"), false);
 
             var o = new
             {
@@ -1487,7 +1487,7 @@ namespace Jint.Tests.Runtime
                         try {
                             throwMyException();
                             return;
-                        } 
+                        }
                         catch(e) {
                             return;
                         }
@@ -1497,7 +1497,7 @@ namespace Jint.Tests.Runtime
                         try {
                             new Thrower().ThrowNotSupportedException();
                             return;
-                        } 
+                        }
                         catch(e) {
                             return;
                         }
@@ -1521,7 +1521,7 @@ namespace Jint.Tests.Runtime
                         try {
                             throwMyException();
                             return '';
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
@@ -1531,7 +1531,7 @@ namespace Jint.Tests.Runtime
                         try {
                             new Thrower().ThrowExceptionWithMessage('myExceptionMessage');
                             return;
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
@@ -1556,7 +1556,7 @@ namespace Jint.Tests.Runtime
                         try {
                             throwMyException1();
                             return '';
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
@@ -1566,7 +1566,7 @@ namespace Jint.Tests.Runtime
                         try {
                             throwMyException2();
                             return '';
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
@@ -1576,7 +1576,7 @@ namespace Jint.Tests.Runtime
                         try {
                             new Thrower().ThrowNotSupportedExceptionWithMessage('myExceptionMessage');
                             return '';
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
@@ -1586,13 +1586,13 @@ namespace Jint.Tests.Runtime
                         try {
                             new Thrower().ThrowArgumentNullException();
                             return '';
-                        } 
+                        }
                         catch(e) {
                             return e.message;
                         }
                     }
                 ");
-            
+
             Assert.Equal(engine.Invoke("throwException1").AsString(), exceptionMessage);
             Assert.Throws<ArgumentNullException>(() => engine.Invoke("throwException2"));
             Assert.Equal(engine.Invoke("throwException3").AsString(), exceptionMessage);