--- a/Jint.Tests/Runtime/JsValueConversionTests.cs
+++ b/Jint.Tests/Runtime/JsValueConversionTests.cs
@@ -3,11 +3,6 @@ using Jint.Native.Array;
 using Jint.Native.Date;
 using Jint.Native.Object;
 using Jint.Native.RegExp;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace Jint.Tests.Runtime
@@ -17,7 +12,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeAnArray()
         {
-            var value = new JsValue(new ArrayInstance(null));
+            var value = new JsObject(new ArrayInstance(null));
             Assert.Equal(false, value.IsBoolean());
             Assert.Equal(true, value.IsArray());
             Assert.Equal(false, value.IsDate());
@@ -35,7 +30,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeABoolean()
         {
-            var value = new JsValue(true);
+            var value = new JsBoolean(true);
             Assert.Equal(true, value.IsBoolean());
             Assert.Equal(false, value.IsArray());
             Assert.Equal(false, value.IsDate());
@@ -53,7 +48,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeADate()
         {
-            var value = new JsValue(new DateInstance(null));
+            var value = new JsObject(new DateInstance(null));
             Assert.Equal(false, value.IsBoolean());
             Assert.Equal(false, value.IsArray());
             Assert.Equal(true, value.IsDate());
@@ -87,7 +82,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeANumber()
         {
-            var value = new JsValue(2);
+            var value = new JsNumber(2);
             Assert.Equal(false, value.IsBoolean());
             Assert.Equal(false, value.IsArray());
             Assert.Equal(false, value.IsDate());
@@ -104,7 +99,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeAnObject()
         {
-            var value = new JsValue(new ObjectInstance(null));
+            var value = new JsObject(new ObjectInstance(null));
             Assert.Equal(false, value.IsBoolean());
             Assert.Equal(false, value.IsArray());
             Assert.Equal(false, value.IsDate());
@@ -121,7 +116,7 @@ namespace Jint.Tests.Runtime
         [Fact]
         public void ShouldBeARegExp()
         {
-            var value = new JsValue(new RegExpInstance(null));
+            var value = new JsObject(new RegExpInstance(null));
             Assert.Equal(false, value.IsBoolean());
             Assert.Equal(false, value.IsArray());
             Assert.Equal(false, value.IsDate());