--- a/Jint.Tests/Runtime/NullPropagation.cs
+++ b/Jint.Tests/Runtime/NullPropagation.cs
@@ -24,12 +24,12 @@ namespace Jint.Tests.Runtime
 
             public bool TryGetCallable(Engine engine, object reference, out JsValue value)
             {
-                value = new JsValue(
+                value = new JsObject(
                     new ClrFunctionInstance(engine, (thisObj, values) => thisObj)
                 );
                 return true;
             }
-            
+
             public bool CheckCoercible(JsValue value)
             {
                 return true;
@@ -42,8 +42,8 @@ namespace Jint.Tests.Runtime
             var engine = new Engine(cfg => cfg.SetReferencesResolver(new NullPropagationReferenceResolver()));
 
             const string Script = @"
-var input = { 
-	Address : null 
+var input = {
+	Address : null
 };
 
 var address = input.Address;