--- a/Jint/Engine.cs
+++ b/Jint/Engine.cs
@@ -67,7 +67,6 @@ namespace Jint
             { typeof(UInt16), (Engine engine, object v) => JsValue.FromInt((UInt16)v) },
             { typeof(UInt32), (Engine engine, object v) => JsValue.FromInt((UInt32)v) },
             { typeof(UInt64), (Engine engine, object v) => JsValue.FromInt((UInt64)v) },
-            { typeof(JsValue), (Engine engine, object v) => (JsValue)v },
             { typeof(System.Text.RegularExpressions.Regex), (Engine engine, object v) => engine.RegExp.Construct((System.Text.RegularExpressions.Regex)v, "") }
         };