--- a/Jint/Native/Date/DateConstructor.cs
+++ b/Jint/Native/Date/DateConstructor.cs
@@ -131,7 +131,7 @@ namespace Jint.Native.Date
                 var v = TypeConverter.ToPrimitive(arguments[0]);
                 if (v.IsString())
                 {
-                    return Construct(Parse(Undefined.Instance, Arguments.From(v)).AsNumber());
+                    return Construct(Parse(Undefined, Arguments.From(v)).AsNumber());
                 }
 
                 return Construct(TypeConverter.ToNumber(v));